requirements.txt
PyYaml
mitmproxy
beautifulsoup4
requests_raw